made content-length support with int64, refactoring, various fixes

+++ common/utilities.go
@@ -5,12 +5,12 @@ import (
 	"os"
 )
 
-func FileSize(fd *os.File) int {
+func FileSize(fd *os.File) int64 {
 	fileinfo, err := fd.Stat()
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	filesize := int(fileinfo.Size())
+	filesize := fileinfo.Size()
 	return filesize
 }
